Move sound_creator module
