+++ src/audio/synth_functions/sound_creator.py
